finished changes in master
