+++ .gitignore
@@ -18,3 +18,4 @@ contract_library
 .pytest_cache
 .DS_Store
 References
+contracts_hold
