+++ khron_brownie/contracts_slither/BasicUseCase.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.8.0;
 
 
 //Local Import from package source
-//import "Khronus_Utils/contracts/KhronusClientBase.sol";
+import "Khronus_Utils/contracts/KhronusClientBase.sol";
 //Remote Import
-import "@khronus/khronus-utils@0.0.3/contracts/KhronusClientBase.sol";
+//import "@khronus/khronus-utils@0.0.3/contracts/KhronusClientBase.sol";
 
 contract BasicClient is KhronusClient{
